Cover the C ABI, and report the right errno when a path is missing
The C ABI had no tests at all. It is the layer a consuming application
actually links against, so a defect there reaches users even though the
Rust-level suite is green — and the sibling drivers have 49 and 17 such
test files, so this crate was the outlier rather than the norm.

One real bug, found by writing them. fs_erofs_mount reported EIO for any
open failure, including a path that simply is not there. A caller
distinguishes "you gave me the wrong path" from "this media is damaged"
only by the errno, so a mistyped path sent a user looking for a hardware
fault. Not-found and permission-denied are now reported as themselves;
everything else stays EIO, which is honest for a device that failed in a
way the caller cannot fix.

The tests cover both read entry points, both alternative mount paths,
NULL in every pointer position, release functions on NULL, iteration
past the end of a directory, reads that cross block boundaries and end
inside a partial one, reads at and past end of file, and a caller
callback that fails — which must surface as an error rather than as
silently zeroed data, since a caller cannot tell those apart.

Two assumptions I got wrong are worth recording. readlink returns 0 on
success rather than a length, and refuses a buffer too small for the
target rather than truncating it. Refusing is the safer choice — a
silently truncated path is a path to somewhere else — but the sibling
XFS and Btrfs drivers truncate instead, so the family is inconsistent
here and should be reconciled deliberately rather than by accident.

Also drops two redundant casts a newer clippy flags in an existing test.

capi.rs coverage 0% -> 91.11% of regions and 92.66% of lines. Every
module in the crate is now above 80%; the crate totals 95.55%.

+++ tests/oracle_compat.rs
@@ -1352,7 +1352,7 @@ fn synthetic_legacy_big_pcluster_two_block_round_trip() {
     img[file_off + 6..file_off + 8].copy_from_slice(&1u16.to_le_bytes()); // nlink
     img[file_off + 8..file_off + 12].copy_from_slice(&(payload.len() as u32).to_le_bytes());
     // raw_blkaddr (i_u for compressed inodes carries `compressed_blocks`).
-    img[file_off + 16..file_off + 20].copy_from_slice(&(blocks as u32).to_le_bytes());
+    img[file_off + 16..file_off + 20].copy_from_slice(&blocks.to_le_bytes());
 
     // ZMap header at file_off + 32 (= body_end for compact 32-byte
     // inode with zero xattrs).
@@ -1382,7 +1382,7 @@ fn synthetic_legacy_big_pcluster_two_block_round_trip() {
                                                                         // delta[0] (low 16): CBLKCNT marker | blocks. delta[1] (high 16):
                                                                         // 3 (forward distance to next HEAD; no next HEAD here, so any
                                                                         // value would do — we pick 3 to mirror what mkfs.erofs emits).
-    let cblkcnt_lo: u32 = 0x0800u32 | (blocks as u32);
+    let cblkcnt_lo: u32 = 0x0800u32 | blocks;
     let u_raw_nh1: u32 = cblkcnt_lo | (3u32 << 16);
     img[nh1_off + 4..nh1_off + 8].copy_from_slice(&u_raw_nh1.to_le_bytes());
     let nh2_off = idx_off + 16;
